Deprecated functions in preload.php
